Pylint fixes (attempt to)
